Split large file into smaller files and added input boxes
